Re-add static_cast in core/Genome

+++ source/core/Genome.hpp
@@ -159,7 +159,7 @@ namespace mabe {
     emp::vector<locus_t> data;                        // Actual data in the genome.
     double mut_p = 0.0;                               // Mutation probability (LOTS TO DO HERE!)
     size_t min_size = 0;
-    size_t max_size = -1;
+    size_t max_size = static_cast<size_t>(-1);
 
     double alphabet_size = 4.0;
